requirements.txt
numpy>=1.21
pandas>=1.3
matplotlib>=3.4
openpyxl>=3.0
pyyaml>=6.0
